Moves FeedStoreSpy to its own file in a Helpers folder within a shared module level scope.
